requirements.txt
aiofile==3.8.8
aiopath==0.6.11
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
Brotli==1.1.0
caio==0.9.13
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dependency-injector==4.41.0
ecdsa==0.19.0
fastapi-slim==0.111.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mutagen==1.47.0
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pyhumps==3.8.0
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-dependency==0.6.0
pytest-order==1.2.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
youtube-search==2.1.2
yt-dlp==2024.4.9
